check for existing user delete/update

+++ Prestige_eSportsAPI/Controllers/UserController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Prestige_eSports.Core.Models;
 using Prestige_eSports.Service.Interfaces;
+using Prestige_eSports.Service.Services;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -69,6 +70,11 @@ namespace Prestige_eSportsAPI.Controllers
             {
                 if (user == null)
                     return BadRequest("User cannot be null");
+
+                var existingUser = _userService.GetFirstOrDefault(user.UserId);
+                if(existingUser == null)
+                    return BadRequest("This user does not exist");
+                    
                 var deleted = await _userService.DeleteUser(user); ;
                 return Ok(deleted);
             }
@@ -85,6 +91,11 @@ namespace Prestige_eSportsAPI.Controllers
             {
                 if (user == null)
                     return BadRequest("User cannot be null");
+
+                var existingUser = _userService.GetFirstOrDefault(user.UserId);
+                if(existingUser == null)
+                    return BadRequest("This user does not exist");
+                
                 var updated = await _userService.UpdateUser(user);
                 return Ok(updated);
             }
